Adding 4 New files
There are 4 new files here.
2 modified and 2 brand new

I'm adding 2 modified files instead of replacing the old ones because some of the chages were drastic and I didn't want to lose the original if you still need/wanted them.

IngSearch.js
is a new javascript file that runs the ingredient search and the switch back and forth between recipe-search and ingredient-search. It is only linked on recipesearch[modified].html

ingSearchStyles.css
is a new style sheet for what was added to recipeSearch[modified].html
it is only linked in recipeSearch[modified].html

recipeSearch[modified].html
is the original search page with additions:
Added radio button to switch between searches
Added hidden classes to the search box for ingredient-search
Added a new container with divs to serve as the result box for both recipe and ingredient search.

recipe[modified].js
is your original js file that I modified to fit the new html.

crossorigin.me has been down for the last 2 days and i've been using a plugin to test this stuff so you may have a hard time seeing the new changes.