small fix to map responsiveness
This isn't a full fix, but it looks better at some lower resolutions but it's still not good on mobil at all. It won't be as long as we use absolute positioning. I think it's good enough as is, everything works and looks good on a full screen, and probably even on some tablets.